// src/test/mutation/client-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  DateString: any;
  FunctionalBoostValue: any;
  Markdown: any;
  NonNegativeInt: any;
  Timestamp: any;
  Url: any;
};
// ...
export type CreateHighlightInput = {
  /** The ID of the Item that should be annotated in the User's list */
  itemId: Scalars['ID'];
  /** Optional note generated by User */
  note?: InputMaybe<Scalars['String']>;
  /**
   * Patch string generated by 'DiffMatchPatch' library, serialized
   * into text via `patch_toText` method.
   * Format is similar to UniDiff but is character-based.
   * The patched text depends on version. For example, the version 2
   * patch surrounds the highlighted text portion with a pair of
   * sentinel tags: '<pkt_tag_annotation></pkt_tag_annotation>'
   * Reference: https://github.com/google/diff-match-patch
   */
  patch: Scalars['String'];
  /**
   * The full text of the highlighted passage. Used as a fallback for
   * rendering highlight if the patch fails.
   */
  quote: Scalars['String'];
  /** Annotation data version */
  version: Scalars['Int'];
};
// ...
export type Highlight = {
  __typename?: 'Highlight';
  /** When the Highlight was created */
  _createdAt: Scalars['Timestamp'];
  /** When the highlight was last updated */
  _updatedAt: Scalars['Timestamp'];
  /** The ID for this Highlight annotation */
  id: Scalars['ID'];
  /** Highlight Note associated with this Highlight */
  note?: Maybe<HighlightNote>;
  /**
   * Patch string generated by 'DiffMatchPatch' library, serialized
   * into text via `patch_toText` method. Use `patch_fromText` to
   * deserialize into an object that can be used by the DiffMatchPatch
   * library. Format is similar to UniDiff but is character-based.
   * The patched text depends on version. For example, the version 2
   * patch surrounds the highlighted text portion with a pair of
   * sentinel tags: '<pkt_tag_annotation></pkt_tag_annotation>'
   * Reference: https://github.com/google/diff-match-patch
   */
  patch: Scalars['String'];
  /**
   * The full text of the highlighted passage. Used as a fallback for
   * rendering highlight if the patch fails.
   */
  quote: Scalars['String'];
  /** Version number for highlight data specification */
  version: Scalars['Int'];
};

export type HighlightNote = {
  __typename?: 'HighlightNote';
  /** When the HighlightNote was created */
  _createdAt: Scalars['Timestamp'];
  /** When the HighlightNote was last updated */
  _updatedAt: Scalars['Timestamp'];
  /** User entered text */
  text: Scalars['String'];
};
// ...
export type Mutation = {
  __typename?: 'Mutation';
  /** Create new highlight note. Returns the data for the created Highlight note. */
  createSavedItemHighlightNote?: Maybe<HighlightNote>;
  /** Create new highlight annotation(s). Returns the data for the created Highlight object(s). */
  createSavedItemHighlights: Array<Highlight>;
  /**
   * Add tags to the savedItems
   * Inputs a list of SavedItemTagsInput(ie. savedItemId and the list of tagName)
   * Returns the list of `SavedItem` for which the tags were added
   */
  createSavedItemTags: Array<SavedItem>;
  /**
   * Deletes a SavedItem from the users list. Returns ID of the
   * deleted SavedItem
   */
  deleteSavedItem: Scalars['ID'];
  /** Delete a highlight by its ID. */
  deleteSavedItemHighlight: Scalars['ID'];
  /** Delete a highlight note by the Highlight ID. */
  deleteSavedItemHighlightNote: Scalars['ID'];
  /**
   * Delete one or more tags from one or more SavedItems.
   * Note that if this operation results in a Tag having no associations
   * to a SavedItem, the Tag object will be deleted.
   */
  deleteSavedItemTags: Array<SavedItem>;
  /**
   * Deletes a Tag object. This is deletes the Tag and all SavedItem associations
   * (removes the Tag from all SavedItems). Returns ID of the deleted Tag.
   */
  deleteTag: Scalars['ID'];
  /** Deletes user information and their pocket data for the given pocket userId. Returns pocket userId. */
  deleteUser: Scalars['ID'];
  /**
   * Deletes user information and their pocket data for the given firefox account ID.
   * Returns firefox account ID sent as the query parameter with the request.
   */
  deleteUserByFxaId: Scalars['ID'];
  /** Refresh an {Item}'s article content. */
  refreshItemArticle: Item;
  /**
   * Replaces the old tags associated with the savedItem to the new tag list
   * given in the entry
   * To remove all Tags from a SavedItem, use `updateSavedItemRemoveTags`.
   * Note: if there is a new tag name in the SavedItemTagsInput, then the tag record will be created
   * Inputs a list of SavedItemTagsInput(ie. savedItemId and list of tag names)
   * Returns the SavedItem for which the tags have been modified.
   */
  replaceSavedItemTags: Array<SavedItem>;
  /** Archives a SavedItem */
  updateSavedItemArchive: SavedItem;
  /** Favorites a SavedItem */
  updateSavedItemFavorite: SavedItem;
  /**
   * Update an existing highlight annotation, by its ID.
   * If the given highlight ID does not exist, will return error data
   * and the highlight will not be created.
   */
  updateSavedItemHighlight: Highlight;
  /**
   * Update an existing highlight note, by its ID.
   * If the given highlight ID does not exist, will return error data
   * and the note will not be updated.
   */
  updateSavedItemHighlightNote?: Maybe<HighlightNote>;
  /**
   * Removes all Tag associations from a SavedItem. Returns the
   * SavedItem that had its Tag associations cleared.
   * Note that if this operation results in a Tag having no associations
   * to a SavedItem, the Tag object will be deleted.
   */
  updateSavedItemRemoveTags: SavedItem;
  /**
   * Set the Tags that are associated with a SavedItem.
   * Will replace any existing Tag associations on the SavedItem.
   * To remove all Tags from a SavedItem, use `updateSavedItemRemoveTags`.
   */
  updateSavedItemTags: SavedItem;
  /** Unarchives a SavedItem */
  updateSavedItemUnArchive: SavedItem;
  /** Undo the delete operation for a SavedItem */
  updateSavedItemUnDelete: SavedItem;
  /** Unfavorites a SavedItem */
  updateSavedItemUnFavorite: SavedItem;
  /**
   * Updates a Tag (renames the tag), and returns the updated Tag.
   * If a Tag with the updated name already exists in the database, will
   * associate that Tag to all relevant SavedItems rather than creating
   * a duplicate Tag object.
   */
  updateTag: Tag;
  /**
   * update the email of the user for the given pocket userId. Request is made by
   * an authenticated user, and the userID is inferred from the request headers `userid`.
   */
  updateUserEmail: User;
  /**
   * update the email of the user for the given firefox account ID. Request
   * is made by a backend service. The `userid` in the headers should match
   * the FxA ID or else an authentication error will be thrown.
   */
  updateUserEmailByFxaId: User;
  /** Updates user preferences for content recommendations across Pocket. */
  updateUserRecommendationPreferences: UserRecommendationPreferences;
  /**
   * Updates a SavedItem, undeletes and unarchives it, bringing it to the top of the user's list, if it exists
   * and creates it if it doesn't.
   */
  upsertSavedItem: SavedItem;
};
// ...
export type SavedItem = RemoteEntity & {
  __typename?: 'SavedItem';
  /** Unix timestamp of when the entity was created */
  _createdAt: Scalars['Int'];
  /** Unix timestamp of when the entity was deleted, 30 days after this date this entity will be HARD deleted from the database and no longer exist */
  _deletedAt?: Maybe<Scalars['Int']>;
  /** Unix timestamp of when the entity was last updated, if any property on the entity is modified this timestamp is set to the modified time */
  _updatedAt?: Maybe<Scalars['Int']>;
  /** Version of the entity, this will increment with each modification of the entity's field */
  _version?: Maybe<Scalars['Int']>;
  /** Annotations associated to this SavedItem */
  annotations?: Maybe<SavedItemAnnotations>;
  /** Timestamp that the SavedItem became archied, null if not archived */
  archivedAt?: Maybe<Scalars['Int']>;
  /** Timestamp that the SavedItem became favorited, null if not favorited */
  favoritedAt?: Maybe<Scalars['Int']>;
  /** Surrogate primary key. This is usually generated by clients, but will be generated by the server if not passed through creation */
  id: Scalars['ID'];
  /** Helper property to indicate if the SavedItem is archived */
  isArchived: Scalars['Boolean'];
  /** Helper property to indicate if the SavedItem is favorited */
  isFavorite: Scalars['Boolean'];
  /** Link to the underlying Pocket Item for the URL */
  item: ItemResult;
  /** The status of this SavedItem */
  status?: Maybe<SavedItemStatus>;
  /** The Suggested Tags associated with this SavedItem, if the user is not premium or there are none, this will be empty. */
  suggestedTags?: Maybe<Array<Tag>>;
  /** The Tags associated with this SavedItem */
  tags?: Maybe<Array<Tag>>;
  /** The url the user saved to their list */
  url: Scalars['String'];
};
// ...
export type UserRecommendationPreferences = {
  __typename?: 'UserRecommendationPreferences';
  /** Topics that the user expressed interest in. */
  preferredTopics?: Maybe<Array<Topic>>;
};
// ...
export type UpdateSavedItemHighlightMutationVariables = Exact<{
  id: Scalars['ID'];
  input: CreateHighlightInput;
}>;


export type UpdateSavedItemHighlightMutation = { __typename?: 'Mutation', updateSavedItemHighlight: { __typename?: 'Highlight', id: string, patch: string, version: number, quote: string, _createdAt: any, _updatedAt: any } };
// ...
export const UpdateSavedItemHighlightDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"updateSavedItemHighlight"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"CreateHighlightInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"updateSavedItemHighlight"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}},{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"patch"}},{"kind":"Field","name":{"kind":"Name","value":"version"}},{"kind":"Field","name":{"kind":"Name","value":"quote"}},{"kind":"Field","name":{"kind":"Name","value":"_createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"_updatedAt"}}]}}]}}]} as unknown as DocumentNode<UpdateSavedItemHighlightMutation, UpdateSavedItemHighlightMutationVariables>;